+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 + [x] Select: add prop `data-placeholder`
 + [x] Select: add prop `data-clear-button` (default: false)
 + [x] Grid: optimise styles
++ [x] Hint: fix remove on leave element
 
 ### 4.2.49
 + [x] Select: extended to interop with role=dropdown